Make Quick optional
