+++ Cartfile
@@ -1,3 +1,2 @@
-github "Quick/Quick"
 github "Quick/Nimble"
 github "facebook/ios-snapshot-test-case" "2.1.4"
